Children of media folder are back to lazy fetch type. Created service for processing folder.
