+++ src/main/kotlin/com/guillermonegrete/gallery/data/MediaFolder.kt
@@ -6,7 +6,7 @@ import javax.persistence.*
 data class MediaFolder(
     @Column(unique = true)
     val name: String = "",
-    @OneToMany(mappedBy = "folder", cascade = [CascadeType.ALL], fetch = FetchType.EAGER)
+    @OneToMany(mappedBy = "folder", cascade = [CascadeType.ALL], fetch = FetchType.LAZY)
     var files: List<MediaFile> = emptyList(),
     @Id @GeneratedValue(strategy = GenerationType.AUTO)
     val id: Long = 0,
